requirements.txt
annotated-types==0.7.0
anyio==4.8.0
Appium-Python-Client==4.5.0
attrs==25.1.0
autogen-agentchat==0.4.3
autogen-core==0.4.3
beautifulsoup4==4.12.3
certifi==2024.12.14
Deprecated==1.2.18
distro==1.9.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
importlib_metadata==8.5.0
jiter==0.8.2
jsonref==1.1.0
numpy==2.2.2
openai==1.60.2
opentelemetry-api==1.29.0
outcome==1.3.0.post0
pandas==2.2.3
pillow==11.1.0
protobuf==4.25.6
pydantic==2.10.6
pydantic_core==2.27.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.2
selenium==4.28.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
tqdm==4.67.1
trio==0.28.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
websocket-client==1.8.0
wrapt==1.17.2
wsproto==1.2.0
zipp==3.21.0
